backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
python-dotenv==1.0.0
Werkzeug==3.0.1
cryptography==41.0.7
Flask-Migrate==4.0.4
